apps/support/requirements.txt
openai>=0.28.0
djangorestframework>=3.14.0 